src/requirements.txt
memory-profiler==0.61.0
google-api-core
google-api-python-client
google.auth
regex
pandas==1.5.2
pandasql
sqlalchemy==1.4.46
py-spy
pyarrow